backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
vosk>=0.3.44
llama-cpp-python>=0.2.0
piper-tts>=1.2.0
